[ORC] Introduce LinkGraphLayer interface and LinkGraphLinkingLayer.
Introduces a new layer interface, LinkGraphLayer, that can be used to add
LinkGraphs to an ExecutionSession.

This patch moves most of ObjectLinkingLayer's functionality into a new
LinkGraphLinkingLayer which should (in the future) be able to be used without
linking libObject. ObjectLinkingLayer now inherits from LinkGraphLinkingLayer
and just handles conversion of object files to LinkGraphs, which are then handed
down to LinkGraphLinkingLayer to be linked.
