+++ llvm/include/llvm/ExecutionEngine/Orc/COFFPlatform.h
@@ -22,6 +22,7 @@
 #include "llvm/ExecutionEngine/Orc/Shared/ExecutorAddress.h"
 
 #include <future>
+#include <list>
 #include <memory>
 #include <thread>
 #include <vector>
